Add files via upload
Most of it were done as training for specific skills which explains the lack of CSS in some of the files. I also gave them different names to simply sort through the files easily. I did not upload one more project here which was the final project for the academy: the message board. I can show that in person if required.